server/requirements.txt
pyqt6-sip
PyQt6==6.9.1
psycopg2-binary
python-dotenv==1.1.1